Move runtime config tests
