+++ runtimeconfig/db_config_test.go
@@ -1,11 +1,10 @@
-package goa4web
+package runtimeconfig
 
 import (
 	"flag"
 	"testing"
 
 	"github.com/arran4/goa4web/config"
-	"github.com/arran4/goa4web/runtimeconfig"
 )
 
 func TestDBConfigPrecedence(t *testing.T) {
@@ -21,7 +20,7 @@ func TestDBConfigPrecedence(t *testing.T) {
 		config.EnvDBPort: "1",
 	}
 	_ = fs.Parse([]string{"--db-pass=cli"})
-	cfg := runtimeconfig.GenerateRuntimeConfig(fs, vals, func(k string) string { return env[k] })
+	cfg := GenerateRuntimeConfig(fs, vals, func(k string) string { return env[k] })
 	if cfg.DBUser != "file" || cfg.DBPass != "cli" || cfg.DBHost != "env" || cfg.DBPort != "1" {
 		t.Fatalf("merged %#v", cfg)
 	}
@@ -32,7 +31,7 @@ func TestLoadDBConfigFromFileValues(t *testing.T) {
 	vals := map[string]string{
 		config.EnvDBUser: "fileval",
 	}
-	cfg := runtimeconfig.GenerateRuntimeConfig(fs, vals, func(string) string { return "" })
+	cfg := GenerateRuntimeConfig(fs, vals, func(string) string { return "" })
 	if cfg.DBUser != "fileval" {
 		t.Fatalf("want fileval got %q", cfg.DBUser)
 	}
